Move all pipelines to training2 directory.
