+++ src/invoke_training/scripts/invoke_dreambooth_lora_sd.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import yaml
 
 from invoke_training.config.pipelines.finetune_lora_config import DreamBoothLoRAConfig
-from invoke_training.training.dreambooth_lora.dreambooth_lora_sd import run_training
+from invoke_training.training2.pipelines.stable_diffusion.dreambooth_lora_sd import run_training
 
 
 def parse_args():
