+++ src/invoke_training/scripts/invoke_textual_inversion_sd.py
@@ -6,7 +6,7 @@ import yaml
 from invoke_training.config.pipelines.textual_inversion_config import (
     TextualInversionConfig,
 )
-from invoke_training.training.textual_inversion.textual_inversion_sd import run_training
+from invoke_training.training2.pipelines.stable_diffusion.textual_inversion_sd import run_training
 
 
 def parse_args():
