+++ src/invoke_training/training2/pipelines/stable_diffusion/finetune_lora_sd.py
@@ -1,26 +1,25 @@
 import json
+import logging
 import math
 import os
 import tempfile
 import time
 
+import numpy as np
 import torch
+import torch.utils.data
+from accelerate import Accelerator
+from accelerate.hooks import remove_hook_from_module
 from accelerate.utils import set_seed
-from diffusers import UNet2DConditionModel
+from diffusers import AutoencoderKL, DDPMScheduler, StableDiffusionPipeline, UNet2DConditionModel
 from diffusers.optimization import get_scheduler
 from tqdm.auto import tqdm
-from transformers import CLIPTextModel
+from transformers import CLIPTextModel, CLIPTokenizer
 
-from invoke_training.config.pipelines.finetune_lora_config import DreamBoothLoRAConfig
+from invoke_training.config.pipelines.finetune_lora_config import FinetuneLoRAConfig
 from invoke_training.core.lora.injection.stable_diffusion import (
     inject_lora_into_clip_text_encoder,
     inject_lora_into_unet,
-)
-from invoke_training.training.finetune_lora.finetune_lora_sd import (
-    cache_vae_outputs,
-    generate_validation_images,
-    load_models,
-    train_forward,
 )
 from invoke_training.training2.shared.accelerator.accelerator_utils import (
     get_mixed_precision_dtype,
@@ -28,12 +27,285 @@ from invoke_training.training2.shared.accelerator.accelerator_utils import (
     initialize_logging,
 )
 from invoke_training.training2.shared.checkpoints.checkpoint_tracker import CheckpointTracker
-from invoke_training.training2.shared.data.data_loaders.dreambooth_sd_dataloader import build_dreambooth_sd_dataloader
+from invoke_training.training2.shared.data.data_loaders.image_caption_sd_dataloader import (
+    build_image_caption_sd_dataloader,
+)
+from invoke_training.training2.shared.data.transforms.tensor_disk_cache import TensorDiskCache
 from invoke_training.training2.shared.optimizer.optimizer_utils import initialize_optimizer
 from invoke_training.training2.shared.stable_diffusion.lora_checkpoint_utils import save_lora_checkpoint
+from invoke_training.training2.shared.stable_diffusion.model_loading_utils import PipelineVersionEnum, load_pipeline
+from invoke_training.training2.shared.stable_diffusion.tokenize_captions import tokenize_captions
 
 
-def run_training(config: DreamBoothLoRAConfig):  # noqa: C901
+def load_models(
+    config: FinetuneLoRAConfig,
+) -> tuple[CLIPTokenizer, DDPMScheduler, CLIPTextModel, AutoencoderKL, UNet2DConditionModel]:
+    """Load all models required for training from disk, transfer them to the
+    target training device and cast their weight dtypes.
+
+    Args:
+        config (FinetuneLoRAConfig): The LoRA training run config.
+        logger (logging.Logger): A logger.
+
+    Returns:
+        tuple[
+            CLIPTokenizer,
+            DDPMScheduler,
+            CLIPTextModel,
+            AutoencoderKL,
+            UNet2DConditionModel,
+        ]: A tuple of loaded models.
+    """
+    pipeline: StableDiffusionPipeline = load_pipeline(config.model, PipelineVersionEnum.SD)
+
+    # Extract sub-models from the pipeline.
+    tokenizer: CLIPTokenizer = pipeline.tokenizer
+    text_encoder: CLIPTextModel = pipeline.text_encoder
+    vae: AutoencoderKL = pipeline.vae
+    unet: UNet2DConditionModel = pipeline.unet
+    noise_scheduler = DDPMScheduler(
+        beta_start=0.00085,
+        beta_end=0.012,
+        beta_schedule="scaled_linear",
+        num_train_timesteps=1000,
+        clip_sample=False,
+        steps_offset=1,
+    )
+
+    # Disable gradient calculation for model weights to save memory.
+    text_encoder.requires_grad_(False)
+    vae.requires_grad_(False)
+    unet.requires_grad_(False)
+
+    # Put models in 'eval' mode.
+    text_encoder.eval()
+    vae.eval()
+    unet.eval()
+
+    return tokenizer, noise_scheduler, text_encoder, vae, unet
+
+
+def cache_text_encoder_outputs(
+    cache_dir: str, config: FinetuneLoRAConfig, tokenizer: CLIPTokenizer, text_encoder: CLIPTextModel
+):
+    """Run the text encoder on all captions in the dataset and cache the results to disk.
+
+    Args:
+        cache_dir (str): The directory where the results will be cached.
+        config (FinetuneLoRAConfig): Training config.
+        tokenizer (CLIPTokenizer): The tokenizer.
+        text_encoder (CLIPTextModel): The text_encoder.
+    """
+    data_loader = build_image_caption_sd_dataloader(config.dataset, batch_size=config.train_batch_size, shuffle=False)
+
+    cache = TensorDiskCache(cache_dir)
+
+    for data_batch in tqdm(data_loader):
+        caption_token_ids = tokenize_captions(tokenizer, data_batch["caption"]).to(text_encoder.device)
+        text_encoder_output_batch = text_encoder(caption_token_ids)[0]
+        # Split batch before caching.
+        for i in range(len(data_batch["id"])):
+            cache.save(data_batch["id"][i], {"text_encoder_output": text_encoder_output_batch[i]})
+
+
+def cache_vae_outputs(cache_dir: str, data_loader: torch.utils.data.DataLoader, vae: AutoencoderKL):
+    """Run the VAE on all images in the dataset and cache the results to disk.
+
+    Args:
+        cache_dir (str): The directory where the results will be cached.
+        data_loader (DataLoader): The data loader.
+        vae (AutoencoderKL): The VAE.
+    """
+    cache = TensorDiskCache(cache_dir)
+
+    for data_batch in tqdm(data_loader):
+        latents = vae.encode(data_batch["image"].to(device=vae.device, dtype=vae.dtype)).latent_dist.sample()
+        latents = latents * vae.config.scaling_factor
+        # Split batch before caching.
+        for i in range(len(data_batch["id"])):
+            cache.save(data_batch["id"][i], {"vae_output": latents[i]})
+
+
+def generate_validation_images(
+    epoch: int,
+    out_dir: str,
+    accelerator: Accelerator,
+    vae: AutoencoderKL,
+    text_encoder: CLIPTextModel,
+    tokenizer: CLIPTokenizer,
+    noise_scheduler: DDPMScheduler,
+    unet: UNet2DConditionModel,
+    config: FinetuneLoRAConfig,
+    logger: logging.Logger,
+):
+    """Generate validation images for the purpose of tracking image generation behaviour on fixed prompts throughout
+    training.
+
+    Args:
+        epoch (int): Epoch number, for reporting purposes.
+        out_dir (str): The output directory where the validation images will be stored.
+        accelerator (Accelerator): Accelerator
+        vae (AutoencoderKL):
+        text_encoder (CLIPTextModel):
+        tokenizer (CLIPTokenizer):
+        noise_scheduler (DDPMScheduler):
+        unet (UNet2DConditionModel):
+        config (FinetuneLoRAConfig): Training configs.
+        logger (logging.Logger): Logger.
+    """
+    logger.info("Generating validation images.")
+
+    # Record original model devices so that we can restore this state after running the pipeline with CPU model
+    # offloading.
+    unet_device = unet.device
+    vae_device = vae.device
+    text_encoder_device = text_encoder.device
+
+    # Create pipeline.
+    pipeline = StableDiffusionPipeline(
+        vae=vae,
+        text_encoder=text_encoder,
+        tokenizer=tokenizer,
+        unet=unet,
+        scheduler=noise_scheduler,
+        safety_checker=None,
+        feature_extractor=None,
+        # TODO(ryand): Add safety checker support.
+        requires_safety_checker=False,
+    )
+    if config.enable_cpu_offload_during_validation:
+        pipeline.enable_model_cpu_offload(accelerator.device.index or 0)
+    else:
+        pipeline = pipeline.to(accelerator.device)
+    pipeline.set_progress_bar_config(disable=True)
+
+    # Run inference.
+    with torch.no_grad():
+        for prompt_idx, prompt in enumerate(config.validation_prompts):
+            generator = torch.Generator(device=accelerator.device)
+            if config.seed is not None:
+                generator = generator.manual_seed(config.seed)
+
+            images = []
+            for _ in range(config.num_validation_images_per_prompt):
+                with accelerator.autocast():
+                    images.append(
+                        pipeline(
+                            prompt,
+                            num_inference_steps=30,
+                            generator=generator,
+                            height=config.dataset.image_transforms.resolution,
+                            width=config.dataset.image_transforms.resolution,
+                        ).images[0]
+                    )
+
+            # Save images to disk.
+            validation_dir = os.path.join(
+                out_dir,
+                "validation",
+                f"epoch_{epoch:0>8}",
+                f"prompt_{prompt_idx:0>4}",
+            )
+            os.makedirs(validation_dir)
+            for image_idx, image in enumerate(images):
+                image.save(os.path.join(validation_dir, f"{image_idx:0>4}.jpg"))
+
+            # Log images to trackers. Currently, only tensorboard is supported.
+            for tracker in accelerator.trackers:
+                if tracker.name == "tensorboard":
+                    np_images = np.stack([np.asarray(img) for img in images])
+                    tracker.writer.add_images(
+                        f"validation (prompt {prompt_idx})",
+                        np_images,
+                        epoch,
+                        dataformats="NHWC",
+                    )
+
+    del pipeline
+    torch.cuda.empty_cache()
+
+    # Remove hooks from models.
+    # HACK(ryand): Hooks get added when calling `pipeline.enable_model_cpu_offload(...)`, but `StableDiffusionPipeline`
+    # does not offer a way to clean them up so we have to do this manually.
+    for model in [unet, vae, text_encoder]:
+        remove_hook_from_module(model)
+
+    # Restore models to original devices.
+    unet.to(unet_device)
+    vae.to(vae_device)
+    text_encoder.to(text_encoder_device)
+
+
+def train_forward(
+    config: FinetuneLoRAConfig,
+    data_batch: dict,
+    vae: AutoencoderKL,
+    noise_scheduler: DDPMScheduler,
+    tokenizer: CLIPTokenizer,
+    text_encoder: CLIPTextModel,
+    unet: UNet2DConditionModel,
+    weight_dtype: torch.dtype,
+) -> torch.Tensor:
+    """Run the forward training pass for a single data_batch.
+
+    Returns:
+        torch.Tensor: Loss
+    """
+    # Convert images to latent space.
+    # The VAE output may have been cached and included in the data_batch. If not, we calculate it here.
+    latents = data_batch.get("vae_output", None)
+    if latents is None:
+        latents = vae.encode(data_batch["image"].to(dtype=weight_dtype)).latent_dist.sample()
+        latents = latents * vae.config.scaling_factor
+
+    # Sample noise that we'll add to the latents.
+    noise = torch.randn_like(latents)
+
+    batch_size = latents.shape[0]
+    # Sample a random timestep for each image.
+    timesteps = torch.randint(
+        0,
+        noise_scheduler.config.num_train_timesteps,
+        (batch_size,),
+        device=latents.device,
+    )
+    timesteps = timesteps.long()
+
+    # Add noise to the latents according to the noise magnitude at each timestep (this is the forward
+    # diffusion process).
+    noisy_latents = noise_scheduler.add_noise(latents, noise, timesteps)
+
+    # Get the text embedding for conditioning.
+    # The text_encoder_output may have been cached and included in the data_batch. If not, we calculate it here.
+    encoder_hidden_states = data_batch.get("text_encoder_output", None)
+    if encoder_hidden_states is None:
+        caption_token_ids = tokenize_captions(tokenizer, data_batch["caption"]).to(text_encoder.device)
+        encoder_hidden_states = text_encoder(caption_token_ids)[0].to(dtype=weight_dtype)
+
+    # Get the target for loss depending on the prediction type.
+    if config.prediction_type is not None:
+        # Set the prediction_type of scheduler if it's defined in config.
+        noise_scheduler.register_to_config(prediction_type=config.prediction_type)
+    if noise_scheduler.config.prediction_type == "epsilon":
+        target = noise
+    elif noise_scheduler.config.prediction_type == "v_prediction":
+        target = noise_scheduler.get_velocity(latents, noise, timesteps)
+    else:
+        raise ValueError(f"Unknown prediction type {noise_scheduler.config.prediction_type}")
+
+    # Predict the noise residual.
+    model_pred = unet(noisy_latents, timesteps, encoder_hidden_states).sample
+
+    loss = torch.nn.functional.mse_loss(model_pred.float(), target.float(), reduction="none")
+    if "loss_weight" in data_batch:
+        # Mean-reduce the loss along all dimensions except for the batch dimension.
+        loss = loss.mean([1, 2, 3])
+        # Apply per-example weights.
+        loss = loss * data_batch["loss_weight"]
+    return loss.mean()
+
+
+def run_training(config: FinetuneLoRAConfig):  # noqa: C901
     # Give a clear error message if an unsupported base model was chosen.
     # TODO(ryan): Update this check to work with single-file SD checkpoints.
     # check_base_model_version(
@@ -78,12 +350,26 @@ def run_training(config: DreamBoothLoRAConfig):  # noqa: C901
         vae.enable_xformers_memory_efficient_attention()
 
     # Prepare text encoder output cache.
-    # text_encoder_output_cache_dir_name = None
+    text_encoder_output_cache_dir_name = None
     if config.cache_text_encoder_outputs:
+        # TODO(ryand): Think about how to better check if it is safe to cache the text encoder outputs. Currently, there
+        # are a number of configurations that would cause variation in the text encoder outputs and should not be used
+        # with caching.
         if config.train_text_encoder:
             raise ValueError("'cache_text_encoder_outputs' and 'train_text_encoder' cannot both be True.")
 
-        raise NotImplementedError("'cache_text_encoder_outputs' is not yet supported in DreamBooth training.")
+        # We use a temporary directory for the cache. The directory will automatically be cleaned up when
+        # tmp_text_encoder_output_cache_dir is destroyed.
+        tmp_text_encoder_output_cache_dir = tempfile.TemporaryDirectory()
+        text_encoder_output_cache_dir_name = tmp_text_encoder_output_cache_dir.name
+        if accelerator.is_local_main_process:
+            # Only the main process should populate the cache.
+            logger.info(f"Generating text encoder output cache ('{text_encoder_output_cache_dir_name}').")
+            text_encoder.to(accelerator.device, dtype=weight_dtype)
+            cache_text_encoder_outputs(text_encoder_output_cache_dir_name, config, tokenizer, text_encoder)
+        # Move the text_encoder back to the CPU, because it is not needed for training.
+        text_encoder.to("cpu")
+        accelerator.wait_for_everyone()
     else:
         text_encoder.to(accelerator.device, dtype=weight_dtype)
 
@@ -100,14 +386,11 @@ def run_training(config: DreamBoothLoRAConfig):  # noqa: C901
         tmp_vae_output_cache_dir = tempfile.TemporaryDirectory()
         vae_output_cache_dir_name = tmp_vae_output_cache_dir.name
         if accelerator.is_local_main_process:
-            # Only the main process should to populate the cache.
+            # Only the main process should populate the cache.
             logger.info(f"Generating VAE output cache ('{vae_output_cache_dir_name}').")
             vae.to(accelerator.device, dtype=weight_dtype)
-            data_loader = build_dreambooth_sd_dataloader(
-                data_loader_config=config.dataset,
-                batch_size=config.train_batch_size,
-                shuffle=False,
-                sequential_batching=True,
+            data_loader = build_image_caption_sd_dataloader(
+                config.dataset, batch_size=config.train_batch_size, shuffle=False
             )
             cache_vae_outputs(vae_output_cache_dir_name, data_loader, vae)
         # Move the VAE back to the CPU, because it is not needed for training.
@@ -156,10 +439,11 @@ def run_training(config: DreamBoothLoRAConfig):  # noqa: C901
 
     optimizer = initialize_optimizer(config.optimizer, trainable_param_groups)
 
-    data_loader = build_dreambooth_sd_dataloader(
-        data_loader_config=config.dataset,
-        batch_size=config.train_batch_size,
-        vae_output_cache_dir=vae_output_cache_dir_name,
+    data_loader = build_image_caption_sd_dataloader(
+        config.dataset,
+        config.train_batch_size,
+        text_encoder_output_cache_dir_name,
+        vae_output_cache_dir_name,
     )
 
     # TODO(ryand): Test in a distributed training environment and more clearly document the rationale for scaling steps
