+++ src/invoke_training/training2/pipelines/stable_diffusion/textual_inversion_sd.py
@@ -15,7 +15,7 @@ from tqdm.auto import tqdm
 from transformers import CLIPTextModel, CLIPTokenizer
 
 from invoke_training.config.pipelines.textual_inversion_config import TextualInversionConfig
-from invoke_training.training.finetune_lora.finetune_lora_sd import (
+from invoke_training.training2.pipelines.stable_diffusion.finetune_lora_sd import (
     cache_vae_outputs,
     generate_validation_images,
     train_forward,
