+++ src/invoke_training/training2/shared/stable_diffusion/tokenize_captions.py
