+++ src/invoke_training/training2/shared/tools/generate_images.py
